mcp_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
pytest==7.4.3
httpx==0.25.2
python-jose[cryptography]==3.3.0
cryptography==41.0.7
alembic==1.13.0
tenacity==8.2.3
ratelimit==2.2.1
prometheus-client==0.19.0
loguru==0.7.2 